Compile would fail with --enable-purify. This patch makes sure it works again. But, how useful is --enable-purify in the current state? We're written a lot of code, but did we account for it to properly work under purify?
git-svn-id: https://svn.apache.org/repos/asf/trafficserver/traffic/trunk@1160602 13f79535-47bb-0310-9956-ffa450edef68

+++ iocore/cluster/ClusterHandler.cc
@@ -1863,7 +1863,7 @@ ClusterHandler::add_small_controlmsg_descriptors()
     c->freeall();
     p = (char *) DOUBLE_ALIGN(p);
 #ifdef PURIFY
-    if ((unsigned int) endp < (unsigned int) p)
+    if ((uintptr_t) endp < (uintptr_t) p)
       memset(endp, 0, (p - endp));
 #endif
   }
